Move getUniqueTopicString from Utils to TestUtils.
