+++ clients/da-vinci-client/src/test/java/com/linkedin/davinci/kafka/consumer/CachedKafkaMetadataGetterTest.java
@@ -13,7 +13,6 @@ import com.linkedin.venice.pubsub.api.PubSubTopic;
 import com.linkedin.venice.pubsub.api.PubSubTopicPartition;
 import com.linkedin.venice.stats.StatsErrorCode;
 import com.linkedin.venice.utils.TestUtils;
-import com.linkedin.venice.utils.Utils;
 import com.linkedin.venice.utils.concurrent.VeniceConcurrentHashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
@@ -65,7 +64,7 @@ public class CachedKafkaMetadataGetterTest {
     CachedKafkaMetadataGetter cachedKafkaMetadataGetter = new CachedKafkaMetadataGetter(1000);
     CachedKafkaMetadataGetter.KafkaMetadataCacheKey key = new CachedKafkaMetadataGetter.KafkaMetadataCacheKey(
         "server",
-        new PubSubTopicPartitionImpl(pubSubTopicRepository.getTopic(Utils.getUniqueTopicString("topic")), 1));
+        new PubSubTopicPartitionImpl(pubSubTopicRepository.getTopic(TestUtils.getUniqueTopicString("topic")), 1));
     Map<CachedKafkaMetadataGetter.KafkaMetadataCacheKey, CachedKafkaMetadataGetter.ValueAndExpiryTime<Long>> offsetCache =
         new VeniceConcurrentHashMap<>();
     CachedKafkaMetadataGetter.ValueAndExpiryTime<Long> valueCache =
