+++ internal/venice-common/src/main/java/com/linkedin/venice/utils/Utils.java
@@ -16,11 +16,9 @@ import com.linkedin.venice.meta.ReadOnlyStoreRepository;
 import com.linkedin.venice.meta.RoutingDataRepository;
 import com.linkedin.venice.meta.Store;
 import com.linkedin.venice.meta.Version;
-import com.linkedin.venice.pubsub.api.PubSubTopicType;
 import com.linkedin.venice.pushmonitor.ExecutionStatus;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.serialization.avro.InternalAvroSpecificSerializer;
-import com.linkedin.venice.views.ChangeCaptureView;
 import java.io.Closeable;
 import java.io.File;
 import java.io.FileInputStream;
@@ -526,29 +524,6 @@ public class Utils {
 
   public static String getUniqueString(String prefix) {
     return String.format("%s_%x_%x", prefix, System.nanoTime(), ThreadLocalRandom.current().nextInt());
-  }
-
-  public static String getUniqueTopicString(String prefix) {
-    int typesNum = PubSubTopicType.values().length;
-    int pubSubTopicTypeIndex = Math.abs(ThreadLocalRandom.current().nextInt() % typesNum);
-    PubSubTopicType pubSubTopicType = PubSubTopicType.values()[pubSubTopicTypeIndex];
-    int version = Math.abs(ThreadLocalRandom.current().nextInt() % typesNum);
-    if (pubSubTopicType.equals(PubSubTopicType.REALTIME_TOPIC)) {
-      return getUniqueString(prefix) + Version.REAL_TIME_TOPIC_SUFFIX;
-    } else if (pubSubTopicType.equals(PubSubTopicType.REPROCESSING_TOPIC)) {
-      return getUniqueString(prefix) + Version.VERSION_SEPARATOR + (version) + Version.STREAM_REPROCESSING_TOPIC_SUFFIX;
-    } else if (pubSubTopicType.equals(PubSubTopicType.VERSION_TOPIC)) {
-      return getUniqueString(prefix) + Version.VERSION_SEPARATOR + (version);
-    } else if (pubSubTopicType.equals(PubSubTopicType.ADMIN_TOPIC)) {
-      return pubSubTopicType.ADMIN_TOPIC_PREFIX + getUniqueString(prefix);
-    } else if (pubSubTopicType.equals(PubSubTopicType.VIEW_TOPIC)) {
-      return getUniqueString(prefix) + Version.VERSION_SEPARATOR + (version)
-          + ChangeCaptureView.CHANGE_CAPTURE_TOPIC_SUFFIX;
-    } else if (pubSubTopicType.equals(PubSubTopicType.UNKNOWN_TYPE_TOPIC)) {
-      return getUniqueString(prefix);
-    } else {
-      throw new VeniceException("Unsupported topic type for: " + pubSubTopicType);
-    }
   }
 
   public static String getUniqueTempPath() {
