+++ internal/venice-test-common/src/integrationtest/java/com/linkedin/venice/writer/VeniceWriterTest.java
@@ -99,7 +99,7 @@ public class VeniceWriterTest {
       int numberOfThreads,
       java.util.function.Consumer<VeniceWriter<KafkaKey, byte[], byte[]>> veniceWriterTask)
       throws ExecutionException, InterruptedException {
-    String topicName = Utils.getUniqueTopicString("topic-for-vw-thread-safety");
+    String topicName = TestUtils.getUniqueTopicString("topic-for-vw-thread-safety");
     int partitionCount = 1;
     PubSubTopic pubSubTopic = pubSubTopicRepository.getTopic(topicName);
     topicManager.createTopic(pubSubTopic, partitionCount, 1, true);
